added _blank target to resources links
