+++ client/src/App.js
@@ -3,7 +3,7 @@ import { BrowserRouter as Router, Route } from "react-router-dom";
 import NavTabs from "./components/NavTabs/NavTabs";
 import Landing from "./pages/jsx/Landing";
 import Video from "./pages/jsx/Video";
-import Resources from "./pages/jsx/Resources.jsx";
+import Resources from "./pages/jsx/Resources";
 import Contact from "./pages/jsx/Contact";
 import Admin from "./pages/jsx/Admin";
 import Footer from "./components/Footer";
